refactor: userドメインのコード改修
