+++ next-app/src/server/service/index.ts
@@ -1,4 +1,1 @@
-export { RegisterService } from "./register.service.js"
-export { LoginStateManagementService } from "./loginStateManagement.service.js"
-export { ProjectService } from "./project.service.js"
 export { UserService } from "./user.service.js"
